Add option to skip generating scope
This adds a `:scopes` option to `boolean_timestamps` that allows the
user to skip generating the scope. This may be desirable in particular
when the generated scope conflicts with an existing Rails method.

For example, `boolean_timestamps :loaded_at` attempts to create a scope
`:loaded` and raises an error because the method `:loaded` is already
defined. Adding the `scopes: false` option avoids the error.

Fixes #7